requirements.txt
setuptools >= 21.0.0
gunicorn==22.0.0
flask-marshmallow==1.1.0
flask-cors==4.0.1
connexion[swagger-ui]==3.1.0
connexion[flask]
connexion[uvicorn]
sqlalchemy==2.0.36
pyjwt==2.4.0
git+https://github.com/kiennt9696/common-utils.git@main
psycopg2==2.9.10
marshmallow_sqlalchemy==1.1.0
alembic==1.14.0